unify transcriber logic in one place
